chatbot-backend/requirements.txt
langchain
langchain-groq
fastapi
uvicorn
python-dotenv
pydantic
pandas